fixed: 패키지 구조 수정
- 도메인 별 패키지 구조 생성
