+++ src/main/java/com/stable/bookrentalsystem/book/client/dto/BookInfo.java
@@ -1,0 +1,16 @@
+package com.stable.bookrentalsystem.book.client.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record BookInfo(
+        String no,
+        @JsonProperty("bk_nm")
+        String title,
+        @JsonProperty("aut_nm")
+        String author,
+        @JsonProperty("pblshr")
+        String publisher,
+        @JsonProperty("pblcn_yr")
+        String publishYear
+) {
+}
