+++ src/main/java/com/stable/bookrentalsystem/book/client/dto/BookInfoResponse.java
@@ -1,0 +1,13 @@
+package com.stable.bookrentalsystem.book.client.dto;
+
+
+import java.util.List;
+
+public record BookInfoResponse(
+        String resultCode,
+        String resultMsg,
+        String numOfRows,
+        String pageNo,
+        String totalCount,
+        List<BookInfo> items
+) {}
