+++ src/main/java/com/stable/bookrentalsystem/book/controller/dto/BookSearchCond.java
@@ -1,0 +1,7 @@
+package com.stable.bookrentalsystem.book.controller.dto;
+
+public record BookSearchCond(
+    String searchTitle,
+    String searchAuthor,
+    String searchPublisher
+) {}
