+++ src/main/java/com/stable/bookrentalsystem/book/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.stable.bookrentalsystem.book.repository;
+
+import com.stable.bookrentalsystem.book.domain.entity.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long>, CustomBookRepository {
+}
