+++ src/main/java/com/stable/bookrentalsystem/book/repository/BulkInsertRepository.java
@@ -1,0 +1,9 @@
+package com.stable.bookrentalsystem.book.repository;
+
+import com.stable.bookrentalsystem.book.domain.entity.Book;
+
+import java.util.List;
+
+public interface BulkInsertRepository {
+    void bulkInsert(List<Book> books);
+}
