--- src/main/java/com/stable/bookrentalsystem/client/book/dto/BookInfo.java
@@ -1,16 +1,0 @@
-package com.stable.bookrentalsystem.client.book.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record BookInfo(
-        String no,
-        @JsonProperty("bk_nm")
-        String title,
-        @JsonProperty("aut_nm")
-        String author,
-        @JsonProperty("pblshr")
-        String publisher,
-        @JsonProperty("pblcn_yr")
-        String publishYear
-) {
-}
