+++ src/main/java/com/stable/bookrentalsystem/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.stable.bookrentalsystem.member.repository;
+
+import com.stable.bookrentalsystem.member.domain.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
