+++ src/main/java/com/stable/bookrentalsystem/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.stable.bookrentalsystem.member.service;
+
+public interface MemberService {
+}
