+++ src/main/java/com/stable/bookrentalsystem/member/service/MemberServiceImpl.java
@@ -1,0 +1,9 @@
+package com.stable.bookrentalsystem.member.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberServiceImpl implements MemberService {
+}
