+++ src/main/java/com/stable/bookrentalsystem/rental/repository/RentalRepository.java
@@ -1,0 +1,7 @@
+package com.stable.bookrentalsystem.rental.repository;
+
+import com.stable.bookrentalsystem.rental.domain.entity.Rental;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RentalRepository extends JpaRepository<Rental, Long> {
+}
