+++ src/main/java/com/stable/bookrentalsystem/rental/service/RentalService.java
@@ -1,0 +1,4 @@
+package com.stable.bookrentalsystem.rental.service;
+
+public interface RentalService {
+}
