+++ src/main/java/com/stable/bookrentalsystem/rental/service/RentalServiceImpl.java
@@ -1,0 +1,9 @@
+package com.stable.bookrentalsystem.rental.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class RentalServiceImpl implements RentalService {
+}
