--- src/main/java/com/stable/bookrentalsystem/repository/book/BulkInsertRepository.java
@@ -1,9 +1,0 @@
-package com.stable.bookrentalsystem.repository.book;
-
-import com.stable.bookrentalsystem.domain.book.entity.Book;
-
-import java.util.List;
-
-public interface BulkInsertRepository {
-    void bulkInsert(List<Book> books);
-}
