--- src/main/java/com/stable/bookrentalsystem/service/member/MemberService.java
@@ -1,4 +1,0 @@
-package com.stable.bookrentalsystem.service.member;
-
-public interface MemberService {
-}
