--- src/main/java/com/stable/bookrentalsystem/service/member/MemberServiceImpl.java
@@ -1,9 +1,0 @@
-package com.stable.bookrentalsystem.service.member;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MemberServiceImpl implements MemberService {
-}
