--- src/main/java/com/stable/bookrentalsystem/service/rental/RentalService.java
@@ -1,4 +1,0 @@
-package com.stable.bookrentalsystem.service.rental;
-
-public interface RentalService {
-}
