--- src/main/java/com/stable/bookrentalsystem/service/rental/RentalServiceImpl.java
@@ -1,9 +1,0 @@
-package com.stable.bookrentalsystem.service.rental;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class RentalServiceImpl implements RentalService {
-}
