--- src/main/java/com/stable/bookrentalsystem/web/dto/book/BookSearchCond.java
@@ -1,7 +1,0 @@
-package com.stable.bookrentalsystem.web.dto.book;
-
-public record BookSearchCond(
-    String searchTitle,
-    String searchAuthor,
-    String searchPublisher
-) {}
